close connections explicitly when done with them
This is important for the PB client because it holds a socket open in a
gen_server.  The link from the webmachine resource to the
riakc_pb_socket is not enough to clean up when the resource finishes,
because it exits with Reason=normal in that case, so the riakc_pb_socket
gen_server lives on.

This adds wrc:disconnect/1, which closes the PB socket, but does nothing
for the HTTP client (since nothing is needed there).

The resources set the 'client' field of their context records or
'disconnected' after closing their client connections, in an effort to
help find bugs where code may try to use a client after it has been
closed.